add cdp browser pools + steel & anchor integrations
